Move types to "dist" directory.
