+++ package.json
@@ -3,7 +3,7 @@
   "version": "4.0.0",
   "description": "Client & Type definitions for the Discord Webhooks API. 📘",
   "main": "dist/index.js",
-  "types": "types/index.d.ts",
+  "types": "dist/index.d.ts",
   "scripts": {
     "test": "tsc --build tsconfig.spec.json",
     "prepublish": "npm test",
@@ -11,8 +11,7 @@
     "prepare": "tsc"
   },
   "files": [
-    "dist/**/*",
-    "types/**/*"
+    "dist/**/*"
   ],
   "repository": {
     "type": "git",
