+++ src/types/Embed/Footer/index.d.ts
