+++ src/types/Embed/Image/index.d.ts
