+++ src/types/Embed/Thumbnail/index.d.ts
